Move LoadTLSCredentials fn to cmd/common.go
